requirements.txt
streamlit==1.32.0
replicate==0.22.0
Pillow==10.2.0
python-dotenv==1.0.1
